requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
httpx==0.27.0
azure-storage-blob==12.19.1
python-multipart==0.0.9
cryptography==42.0.5
